[#1428] fixes navigation issues when adding a new map
+also extract navigation code to navigator class
