+++ offlinemaps/src/main/java/org/akvo/flow/offlinemaps/Constants.java
@@ -25,4 +25,8 @@ public class Constants {
 
     public static final String MAPBOX_MAP_STYLE = Style.MAPBOX_STREETS;
     public static final int MAP_BOX_ZOOM_MAX = 2;
+
+    public static final String CALLING_SCREEN_EXTRA = "calling-screen-extra";
+    public static final int CALLING_SCREEN_EXTRA_DIALOG = 0;
+    public static final int CALLING_SCREEN_EXTRA_LIST = 1;
 }
